requirements.txt
# Core Dependencies
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-generativeai==0.3.1

# Data Processing
openpyxl==3.1.2
pandas==2.1.4
numpy==1.26.2

# Git Operations
gitpython==3.1.40

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
pydantic==2.5.2

# Retry Logic
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
